refactor(PrintConstant): 상수 수정 및 클래스명 변경
